requirements.txt
# Telegram Bot
python-telegram-bot==20.3

# Environment
python-dotenv>=1.0.0

# HTTP requests and web scraping
requests>=2.28.0
aiohttp>=3.8.0

# Optional: AI analysis (uncomment if using OpenAI)
# openai>=1.0.0

# Optional: Enhanced fuzzy matching (uncomment if needed)
# fuzzywuzzy>=0.18.0
# python-Levenshtein>=0.20.0
